feat: add better startup messages to cli
