+++ crates/glaredb/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod commands;
 pub mod local;
 pub mod metastore;
 pub mod pg_proxy;
@@ -5,5 +6,6 @@ pub mod rpc_proxy;
 pub mod server;
 pub mod util;
 
+pub mod args;
 mod highlighter;
 mod prompt;
